Move filter storage into ui bundle
