+++ src/Sylius/Bundle/AdminBundle/spec/Controller/RedirectHandlerSpec.php
@@ -14,9 +14,9 @@ declare(strict_types=1);
 namespace spec\Sylius\Bundle\AdminBundle\Controller;
 
 use PhpSpec\ObjectBehavior;
-use Sylius\Bundle\AdminBundle\Storage\FilterStorageInterface;
 use Sylius\Bundle\ResourceBundle\Controller\RedirectHandlerInterface;
 use Sylius\Bundle\ResourceBundle\Controller\RequestConfiguration;
+use Sylius\Bundle\UiBundle\Storage\FilterStorageInterface;
 use Sylius\Component\Resource\Model\ResourceInterface;
 use Symfony\Component\HttpFoundation\Response;
