+++ src/Sylius/Bundle/UiBundle/Storage/FilterStorage.php
@@ -11,7 +11,7 @@
 
 declare(strict_types=1);
 
-namespace Sylius\Bundle\AdminBundle\Storage;
+namespace Sylius\Bundle\UiBundle\Storage;
 
 use Symfony\Component\HttpFoundation\RequestStack;
 use Symfony\Component\HttpFoundation\Session\SessionInterface;
