+++ src/Sylius/Bundle/UiBundle/Storage/FilterStorageInterface.php
@@ -11,7 +11,7 @@
 
 declare(strict_types=1);
 
-namespace Sylius\Bundle\AdminBundle\Storage;
+namespace Sylius\Bundle\UiBundle\Storage;
 
 interface FilterStorageInterface
 {
